Fix out of the box linting
Running `npm run lint` on a freshly cloned copy of the project failed
because it did not find the correct prettier configuration.

Running `npm run lint-init` made the linting succeed but it also meant
that we have to always run `npm run lint-init` before linting to make
it reliable. The better solution is to make sure that prettier finds
the correct configuration by itself.