Fix: Missing release notes after app update
After updating an app the others were not showing release notes anymore.

This was caused by this: After we update an app, we reload the data for
all online available apps (the function is called `getOfficialApps` but
it really loads from all sources). This causes existing data about the
apps to be wiped, including the release notes. But as an optimisation
we reload only the release notes about the just updated app, so the
other apps are then left without release notes.

This fix does not wipe the existing data about all apps, only data about
the just updated app is overwritten with the online data about it, thus
the release notes from other apps are not removed anymore.